requirements.txt
request==2.28.0
pyyaml==5.4.1
telethon==1.25.4
numpy==1.23.2
nltk==3.7
sklearn==0.0
tflearn==0.5.0